Refactor weekly goal handling
This commit refactors the weekly goal feature by making the following changes:

-   Updates the `WeeklyGoalEntity` to use a fixed primary key instead of an auto-generated one, ensuring only one weekly goal exists at a time.
-   Modifies the goal setting logic in `SetWeeklyTargetViewModel` to overwrite the existing goal instead of deleting and creating a new one.
-   Enhances the goal reset functionality in `GoalProgressViewModel`:
    -   If there are no completed problems, the goal is deleted.
    -   Otherwise, the goal is updated with the remaining uncompleted problems.
-   Improves the `GoalProgressScreen` UI to show a message when no goal is set and prevents resetting an empty goal.
-   Optimizes filtering logic in `GoalProgressViewModel`.

+++ app/src/main/kotlin/com/byteutility/dev/leetcode/plus/ui/screens/targetstatus/GoalProgressViewModel.kt
@@ -1,6 +1,5 @@
 package com.byteutility.dev.leetcode.plus.ui.screens.targetstatus
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.byteutility.dev.leetcode.plus.data.model.LeetCodeProblem
@@ -98,13 +97,12 @@ class GoalProgressViewModel @Inject constructor(
         val goalStartDate =
             goalEntity?.toWeeklyGoalPeriod()?.startDate
         if (goalStartDate != null) {
+            val goalProblemsTitleList = goalEntity.toProblems().map { it.titleSlug }
             val parsedGoalStartDateTime = LocalDate.parse(goalStartDate, formatter1).atStartOfDay()
             return userSubmissions.filter { submission ->
                 val submissionDateTime =
                     LocalDateTime.parse(submission.timestamp, formatter2)
-                val isIncludedInGoal = goalEntity?.toProblems()
-                    ?.map { it.titleSlug }
-                    ?.contains(submission.titleSlug) == true
+                val isIncludedInGoal = goalProblemsTitleList.contains(submission.titleSlug)
                 isIncludedInGoal && submissionDateTime.isAfter(parsedGoalStartDateTime)
             }
         } else {
@@ -130,10 +128,14 @@ class GoalProgressViewModel @Inject constructor(
                         completedProblemsWithStatus.any { goal.title == it.title }
                     }
 
-
-                Log.e(this.javaClass.name, "resetGoal: completedProblems $completedProblems")
-                goalRepository.deleteWeeklyGoal()
-                goalRepository.saveWeeklyGoal(completedProblems, goalProblems.toWeeklyGoalPeriod())
+                if (completedProblems.isNotEmpty()) {
+                    goalRepository.saveWeeklyGoal(
+                        problems = completedProblems,
+                        period = goalProblems.toWeeklyGoalPeriod()
+                    )
+                } else {
+                    goalRepository.deleteWeeklyGoal()
+                }
             }
         }
     }
